fix: lock thread in coinitService
coinitService implicitly required that runtime.LockOSThread() was
called before (because ole.CoInitializeEx requires that all later
calls happen on the same thread). This was true for one location
that called coinitService, but not for another, causing WMI
failures if the go runtime moved the go routine to another thread
during the WMI calls.
To fix this, explicitly call runtime.LockOSThread() in coinitService
and runtime.UnlockOSThread() in the cleanup.